[UPDATE] testcode

+++ src/main/java/com/dna/backend/DNABackend/service/timeline/TimelineServiceImpl.java
@@ -8,9 +8,12 @@ import com.dna.backend.DNABackend.entity.timeline.enums.Type;
 import com.dna.backend.DNABackend.entity.user.User;
 import com.dna.backend.DNABackend.entity.user.UserRepository;
 import com.dna.backend.DNABackend.exception.TimelineNotFoundException;
+import com.dna.backend.DNABackend.exception.UserNotFoundException;
 import com.dna.backend.DNABackend.payload.request.TimelineRequest;
+import com.dna.backend.DNABackend.payload.response.LackOfPermissionException;
 import com.dna.backend.DNABackend.payload.response.TimelineListResponse;
 import com.dna.backend.DNABackend.payload.response.TimelineResponse;
+import com.dna.backend.DNABackend.security.jwt.auth.AuthenticationFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -27,25 +30,29 @@ public class TimelineServiceImpl implements TimelineService{
     private final TimelineRepository timelineRepository;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
+    private final AuthenticationFacade authenticationFacade;
 
 
     @Override
     public void createTimeline(TimelineRequest timelineRequest) {
-        // 대충 토큰검사 해
-        User testUser = userRepository.findById("email2@dsm.hs.kr").get();
+        if(!authenticationFacade.isLogin()) {
+            throw new UserNotFoundException();
+        }
 
-        timelineRepository.save(timelineRequest.toEntity(testUser));
+        User user = userRepository.findById(authenticationFacade.getUserEmail())
+                .orElseThrow(UserNotFoundException::new);
+
+        timelineRepository.save(timelineRequest.toEntity(user));
 
     }
 
     @Override
     public TimelineListResponse getTimelines(Type type, Pageable page) {
-
-        User tmpUser = User.builder()
-                .email("test@dsm.hs.kr")
-                .password("password")
-                .name("hong!")
-                .build();
+        User user = null;
+        if(authenticationFacade.isLogin()) {
+            user = userRepository.findByEmail(authenticationFacade.getUserEmail())
+                    .orElseThrow(UserNotFoundException::new);
+        }
 
         Page<Timeline> timelines;
 
@@ -65,7 +72,7 @@ public class TimelineServiceImpl implements TimelineService{
                             .name(timeline.getUser().getName())
                             .createdAt(timeline.getCreatedAt())
                             .type(timeline.getType())
-                            .isMine(timeline.getUser().equals(tmpUser)) // 토큰 되면 진짜 유저가 올거야
+                            .isMine(timeline.getUser().equals(user))
                             .build()
             );
         }
@@ -80,9 +87,19 @@ public class TimelineServiceImpl implements TimelineService{
 
     @Override
     public void deleteTimeline(Long timelineId) {
-        // 대충 토큰검사 & 본인인지 확인
+        if(!authenticationFacade.isLogin()) {
+            throw new UserNotFoundException();
+        }
+
+        User user = userRepository.findById(authenticationFacade.getUserEmail())
+                .orElseThrow(UserNotFoundException::new);
+
         Timeline timeline = timelineRepository.findById(timelineId)
                 .orElseThrow(TimelineNotFoundException::new);
+
+        if(!user.equals(timeline.getUser())) {
+            throw new LackOfPermissionException();
+        }
 
         for(Comment comment : timeline.getCommentList()) {
             commentRepository.delete(comment);
